Get KeyNav tabIndex working.
Don't put tabIndex in the KeyNav prototype.  It triggers Stateful to create a custom
tabIndex setter which then triggers Widget#enteredViewCallback to
call the setter on widget startup.  We don't want any of these things to happen.
Rather, the app should be able to set/change tabIndex on the element (i.e. widget
root node) without any interference/action from delite.

The only thing that we really need to do is to set tabIndex to 0 if it was not
specified explicitly by the app.

Fixes #62.