Experiment to use jQuery's Deferred rather than dojo's.
It doesn't work currently because $.when() has weird behavior.
when() is supposed to branch depending on whether or not the argument is a promise.
However, the check it actually does is jQuery.isFunction( subordinate.promise ), where
subordinate is the argument, so it doesn't work if you pass in a promise, (there's no
promise.promise() function), nor even if you pass in a normal Deferred (since Deferred#promise
is an object, not a function).